Prevent vcpkg from injecting nonsense into the linker commandline
Looks like the vcpkg integration for Visual Studio will blindly ask the linker to include every installed vcpkg package, by silently injecting into the linker command-line. At least in the case I tested where there are no packages present (and possibly the case where the *are* packages present), this results in a glob which Crinkler falls over on. So, I'm disabling vcpkg at the project level to prevent this.

Thanks to @totetmatt for the initial bug report.